Extracted computer to its own component
